# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AI-Powered Developer Assistant API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /repositories/{repoId}/pull-requests/{prId}/review:
    get:
      operationId: getPullRequestReview
      tags: [reviews]
      summary: Get AI review for a pull request
      parameters:
        - name: repoId
          in: path
          required: true
          schema:
            type: integer
        - name: prId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: AI review result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AIReview"
    post:
      operationId: triggerPullRequestReview
      tags: [reviews]
      summary: Trigger AI review for a pull request
      parameters:
        - name: repoId
          in: path
          required: true
          schema:
            type: integer
        - name: prId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "202":
          description: Review triggered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AIReview"
  /repositories/{repoId}/pull-requests/{prId}/risk:
    get:
      operationId: getPullRequestRisk
      tags: [reviews]
      summary: Get risk score for a pull request
      parameters:
        - name: repoId
          in: path
          required: true
          schema:
            type: integer
        - name: prId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Risk assessment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RiskAssessment"
  /reviews:
    get:
      operationId: listReviews
      tags: [reviews]
      summary: List AI reviews with filtering and pagination
      parameters:
        - name: repoId
          in: query
          schema:
            type: integer
        - name: risk
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of AI reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AIReview"
  /reviews/{id}:
    get:
      operationId: getReview
      tags: [reviews]
      summary: Get AI review details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: AI review details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AIReview"
    delete:
      operationId: deleteReview
      tags: [reviews]
      summary: Delete an AI review
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Review deleted
  /repositories/{repoId}/review:
    post:
      operationId: triggerRepositoryReview
      tags: [reviews]
      summary: Trigger AI review for a repository
      parameters:
        - name: repoId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "202":
          description: Review triggered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AIReview"
...
components:
  schemas:
...
    AIReview:
      type: object
      properties:
        id:
          type: integer
        prId:
          type: integer
        repositoryId:
          type: integer
        repositoryName:
          type: string
        pullRequestNumber:
          type: integer
        branch:
          type: string
        fileName:
          type: string
        userId:
          type: integer
        status:
          type: string
          enum: [pending, in_progress, completed, failed]
        summary:
          type: string
        reviewText:
          type: string
        bugsFound:
          type: integer
        bugs:
          type: array
          items:
            $ref: "#/components/schemas/ReviewIssue"
        improvements:
          type: array
          items:
            $ref: "#/components/schemas/ReviewIssue"
        performance:
          type: array
          items:
            $ref: "#/components/schemas/ReviewIssue"
        refactoring:
          type: array
          items:
            $ref: "#/components/schemas/ReviewIssue"
        securityIssues:
          type: array
          items:
            $ref: "#/components/schemas/ReviewIssue"
        complexity:
          type: string
        riskScore:
          type: integer
        reviewDuration:
          type: integer
        tokensUsed:
          type: integer
        overallScore:
          type: integer
          minimum: 0
          maximum: 100
        createdAt:
          type: string
          format: date-time
      required: [id, status]
    ReviewIssue:
      type: object
      properties:
        severity:
          type: string
          enum: [info, warning, error]
        message:
          type: string
        file:
          type: string
        line:
          type: integer
        suggestion:
          type: string
      required: [severity, message]
    RiskAssessment:
      type: object
      properties:
        prId:
          type: integer
        riskScore:
          type: number
          minimum: 0
          maximum: 100
        riskLevel:
          type: string
          enum: [low, medium, high]
        factors:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              value:
                type: number
              weight:
                type: number
            required: [name, value, weight]
        prediction:
          type: string
      required: [prId, riskScore, riskLevel, factors, prediction]
...